user-service/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.6
PyJWT==2.8.0
Werkzeug==2.3.7
pytest
pytest-cov